Sang_23-08

+++ src/main/java/DoUong_HoaDon_ThongKe_Repository/iThongKe.java
@@ -13,13 +13,13 @@ import java.util.ArrayList;
  */
 public interface iThongKe {
     
-    double getTongDoanhThu();
+    long getTongDoanhThu();
     
     int getTongHoaDon();
     
     int getTongSanPham();
     
-    double getTongDonhThuTheoNgayChon(String d1, String d2);
+    long getTongDonhThuTheoNgayChon(String d1, String d2);
     
     int getTongHoaDonTheoNgayChon(String d1, String d2);
